net:openvswitch: Properly report monitoring drops
If a flow ends with a userspace() or sample(userspace()) action and the
userspace action is of monitoring nature, it means that no futher
actions will be applied to this packet and therefore, it's being
dropped.

Report the packet drop via the right ovs_kfree_skb_reason in this case.

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>